Finish category and ad creation

+++ database/migrations/2019_10_08_193336_create_ads_table.php
@@ -19,7 +19,7 @@ class CreateAdsTable extends Migration
             $table->boolean('approved')->default(false);
             $table->integer('category');
             $table->string('name');
-            $table->float('price');
+            $table->decimal('price', 12, 2);
             $table->integer('year');
             $table->float('range');
             $table->string('image');
